pyproject.toml
[project]
name = "ai_real_image_classification"
version = "0.0.1"
description = "Classification system able to identify AI-generated images (synthetic) vs. authentic photography (real)."
authors = [
  { name = "DTU 02476 Machine learning operations Group 75", email = "your@email.com" },
]
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "datasets>=4.5.0",
  "fastapi==0.115.12",
  "hydra-core>=1.3.2",
  "kagglehub[pandas-dataset]",
  "lightning>=2.6.0",
  "omegaconf>=2.3.0",
  "pandas>=2.3.3",
  "pillow>=12.1.0",
  "scikit-learn>=1.8.0",
  "torch==2.6.0",
  "torchvision>=0.21.0",
  "typer==0.15.2",
  "uvicorn==0.34.2",
  "wandb>=0.24.0",
]

[project.optional-dependencies]
dev = [
  # contents of requirements_dev.txt
]

[dependency-groups]
dev = [
    "coverage==7.8.0",
    "invoke==2.2.0",
    "mkdocs==1.6.1",
    "mkdocs-material==9.6.1",
    "mkdocstrings-python==1.16.10",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff>=0.14.11",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[tool.uv]
package = true
